Add Wiznet W5100 to menuconfig
Wiznet W5100 is the ethernet chip used on the Arduino
Ethernet Shield.  Actually it has a full TCP/IP stack
on board, but it can be used in Raw-MAC-mode.

For Ethersex it is an alternative to ENC28J60, currently
it's not a goal to have both available simultaneously.
Hence the ENC28J60 IP/MAC-configuration is simply re-used.